Added : Ability to trash particular note by ID only for authenticated user

+++ src/services/note.service.js
@@ -7,35 +7,53 @@ export const createNote = async (body) => {
 };
 
 //get all notes
-export const getAllNotes = async () => {
-    const data = await Note.find();
-    return data;
+export const getAllNotes = async (userID) => {
+    const data = await Note.find({userID});
+    if (data.length != 0) {
+      return data;
+    }
+    else {
+      throw new Error("Notes are not created for this user");
+    }
   };
 
 //get note by _id
-export const getNoteByID = async (_id) => {
-    const data = await Note.findById(_id);
-    return data;
+export const getNoteByID = async (_id,userID) => {
+    const data = await Note.findOne({_id,userID});
+    if (data!== null) {
+      return data;
+    } else {
+      throw new Error("Incorrect note ID");
+    }
   };
 
   //update note by _id
-export const updateNoteByID = async (_id, body) => {
+export const updateNoteByID = async (_id,userID, body) => {
     const data = await Note.findOneAndUpdate(
       {
-        _id
+        _id,
+        userID
       },
       body,
       {
         new: true
       }
     );
-    return data;
+    if (data!== null) {
+      return data;
+    } else {
+      throw new Error("Incorrect note ID");
+    }
   };
 
   //delete note by _id
-export const deleteNoteByID = async (_id) => {
-    const data = await Note.findOneAndDelete(_id);
-    return data;
+export const deleteNoteByID = async (_id,userID) => {
+    const data = await Note.findOneAndDelete({_id,userID});
+    if (data!== null) {
+      return data;
+    } else {
+      throw new Error("Incorrect note ID");
+    }
   };
   
   //Archive note by _id
